requirements.txt
py-cord~=2.5.0
tortoise-orm~=0.19.3
python-dotenv~=1.0.0
matplotlib~=3.8.3
pandas~=2.2.1